Allow to run the container with any User ID
In the last commit we created a non-root user "build" to be
used for any operations run in the container. However, we
assumed that User ID of the user is 1000 which doesn't
have to be true.
Switch to a different approach in which the container starts
as a root and takes RUNAS environmental variable to specify
which User ID it should run with. The entrypoint script then
creates a user with the User ID ans switches to the user
with "su" command.